output consistency: statistics: split and move files
