+++ misc/python/materialize/output_consistency/status/consistency_test_logger.py
@@ -1,0 +1,18 @@
+# Copyright Materialize, Inc. and contributors. All rights reserved.
+#
+# Use of this software is governed by the Business Source License
+# included in the LICENSE file at the root of this repository.
+#
+# As of the Change Date specified in that file, in accordance with
+# the Business Source License, use of this software will be governed
+# by the Apache License, Version 2.0.
+
+
+class ConsistencyTestLogger:
+    def __init__(
+        self,
+    ) -> None:
+        self.global_warnings: list[str] = []
+
+    def add_global_warning(self, message: str) -> None:
+        self.global_warnings.append(message)
